Rel 4.10.0.0

+++ src/soc/nvidia/tegra210/mmu_operations.c
@@ -17,13 +17,12 @@
 #include <assert.h>
 #include <soc/addressmap.h>
 #include <soc/mmu_operations.h>
-#include <stdlib.h>
-#include <stdint.h>
 #include <symbols.h>
+#include <types.h>
 
 static void tegra210_mmu_config(void)
 {
-	uint64_t start,end;
+	uintptr_t start, end;
 	const unsigned long devmem = MA_DEV | MA_S | MA_RW;
 	const unsigned long cachedmem = MA_MEM | MA_NS | MA_RW;
 	const unsigned long secure_mem = MA_MEM | MA_S | MA_RW;
